refactor: move labels to the api package; add ScanJob name label
