+++ api/labels.go
@@ -1,8 +1,9 @@
-package handlers
+package api
 
 const (
 	LabelManagedByKey   = "app.kubernetes.io/managed-by"
 	LabelManagedByValue = "sbombastic"
 	LabelPartOfKey      = "app.kubernetes.io/part-of"
 	LabelPartOfValue    = "sbombastic"
+	LabelScanJob        = "sbombastic.rancher.io/scanjob"
 )
